+++ libs/plugin-keyvault/tests/KeyStoreKeyVault.spec.ts
@@ -46,11 +46,11 @@ afterEach(() => {
 describe('KeyStoreKeyVault', () => {
   const alg = { name: 'ECDSA', namedCurve: 'SECP256K1', hash: { name: 'SHA-256' } };
   if (!keyVaultEnable) {
-    console.log('Key vault is enabled. Add your credentials to Credentials.ts')
+    console.log('Key vault is not enabled. Add your credentials to Credentials.ts')
     return;
   }
 
-  it('should create an instance', () =>{
+  it('should create an instance', () => {
     const cache = new KeyStoreInMemory();
     const credential = new ClientSecretCredential(tenantId, clientId, clientSecret);
     let vault = 'https://example.keyvault.com';
@@ -74,6 +74,7 @@ describe('KeyStoreKeyVault', () => {
       expect(list[name]).toBeDefined();
       // Two requests should hit cache
       let key = await keyStore.get(new KeyReference(name, 'key'), new KeyStoreOptions({ latestVersion: false }));
+      key = await keyStore.get(new KeyReference(name, 'key'), new KeyStoreOptions({ latestVersion: false }));
       expect(key).toBeDefined();
       expect((await cache.list())[name]).toBeDefined();
     } finally {
@@ -88,11 +89,11 @@ describe('KeyStoreKeyVault', () => {
     const keyStore = new KeyStoreKeyVault(credential, vaultUri, cache);
     try {
       const secret1 = base64url.encode(name);
-      const secret2 = base64url.encode(name+'2');
+      const secret2 = base64url.encode(name + '2');
 
       //save two versions
-      await keyStore.save( new KeyReference(name), secret1);
-      await keyStore.save( new KeyReference(name), secret2);
+      await keyStore.save(new KeyReference(name), secret1);
+      await keyStore.save(new KeyReference(name), secret2);
 
       let list = await keyStore.list('secret', new KeyStoreOptions({ latestVersion: false }));
       expect(list[name]).toBeDefined();
@@ -111,6 +112,7 @@ describe('KeyStoreKeyVault', () => {
     }
   });
   it('should list a named stored secret with EC', async () => {
+    let cleaned = false;
     const name = 'KvTest-KeyStoreKeyVault' + Math.random().toString(10).substr(2);
     const cache = new KeyStoreInMemory();
     const credential = new ClientSecretCredential(tenantId, clientId, clientSecret);
@@ -122,18 +124,19 @@ describe('KeyStoreKeyVault', () => {
       const jwk: any = await subtle.exportKey('jwk', keyPair.privateKey);
 
       //save two versions
-      await keyStore.save( new KeyReference(name), jwk);
+      await keyStore.save(new KeyReference(name), jwk);
 
       // save okp version
       let okp = clone(jwk);
       okp.kty = 'OKP';
-      await keyStore.save( new KeyReference(name), jwk);
+      await keyStore.save(new KeyReference(name), jwk);
 
       let list = await keyStore.list('secret', new KeyStoreOptions({ latestVersion: false }));
       expect(list[name]).toBeDefined();
 
-      // get latest version only
+      // get latest version only, second should 
       let key = await keyStore.get(new KeyReference(name, 'secret'), new KeyStoreOptions({ latestVersion: true }));
+      key = await keyStore.get(new KeyReference(name, 'secret'), new KeyStoreOptions({ latestVersion: true }));
       expect(key.keys.length).toEqual(1);
       //TODO BUG. k is reported as object
       expect((await cache.list())[name]).toBeDefined();
@@ -147,17 +150,22 @@ describe('KeyStoreKeyVault', () => {
       expect(key.keys.length).toEqual(1);
 
       // negative cases
-      /** ROB TODO
-      getKeyStoreClientSpy: jasmine.Spy = spyOn(keyStore, 'getKeyStoreClient').and.callFake(() => () => {
-        return {
-          keys: {
-            
-          }
-        }
-      });      
-      */
+      cleaned = true;
+        await (<SecretClient>keyStore.getKeyStoreClient('secret')).beginDeleteSecret(name);
+        const getKeyStoreClientSpy: jasmine.Spy = spyOn(keyStore, 'getKeyStoreClient').and.callFake(() => {
+          throw new Error('some error');
+      });         
+      try {
+        await keyStore.get(new KeyReference(name, 'secret'), new KeyStoreOptions({ latestVersion: true }));        
+        fail('get should have thrown');
+      } catch (exception) {
+        expect(exception.message).toEqual('some error');
+        
+      }  
     } finally {
-      await (<SecretClient>keyStore.getKeyStoreClient('secret')).beginDeleteSecret(name);
+      if (!cleaned) {
+        await (<SecretClient>keyStore.getKeyStoreClient('secret')).beginDeleteSecret(name);
+      }
     }
   });
 
@@ -173,11 +181,11 @@ describe('KeyStoreKeyVault', () => {
       await provider.onGenerateKey(alg, false, ['sign'], { keyReference: new KeyReference(name) });
 
       let parts = (<any>keyPair.publicKey.algorithm).kid.split('/');
-      const keyName = `${parts[parts.length-2]}/${parts[parts.length-1]}`;
+      const keyName = `${parts[parts.length - 2]}/${parts[parts.length - 1]}`;
 
       let list = await keyStore.list('key', new KeyStoreOptions({ latestVersion: false }));
       expect(list[name].kids[0].includes(keyName) || list[name].kids[1].includes(keyName)).toBeTruthy();
-      const key = await keyStore.get(new KeyReference(name, 'key',keyName), new KeyStoreOptions({ latestVersion: false }));
+      const key = await keyStore.get(new KeyReference(name, 'key', keyName), new KeyStoreOptions({ latestVersion: false }));
       expect(key.keys.length).toEqual(1);
       console.log(`name: ${keyName}`);
       console.log(`${JSON.stringify(key.keys[0])}`);
@@ -209,17 +217,20 @@ describe('KeyStoreKeyVault', () => {
     const cache = new KeyStoreInMemory();
     const credential = new ClientSecretCredential(tenantId, clientId, clientSecret);
     const keyStore = new KeyStoreKeyVault(credential, vaultUri, cache);
-    let throwed = false;
+    await keyStore.save(new KeyReference(name, 'secret'), 'abcdefg');
+    let list = await keyStore.list('secret', new KeyStoreOptions({ latestVersion: false }));
+    expect(list[name]).toBeDefined();
     try {
-      await keyStore.save(new KeyReference(name, 'secret'), 'abcdefg');
-      let list = await keyStore.list('secret', new KeyStoreOptions({ latestVersion: false }));
-      expect(list[name]).toBeDefined();
       await cache.get(new KeyReference(name, 'secret'));
-      expect(throwed).toBeTruthy();
+      fail('Should have thrown during get: should set a secret');
     } catch (err) {
-      throwed = true;
       expect(err.message).toEqual(`${name} not found`)
-
+    } 
+    try {
+      await keyStore.save(<any>undefined, '');
+      fail('Should have thrown during save: should set a secret');
+    } catch (err) {
+      expect(err.message).toEqual(`Key reference needs to be specified`)
     } finally {
       await (<SecretClient>keyStore.getKeyStoreClient('secret')).beginDeleteSecret(name);
     }