+++ CHANGE_LOG.md
@@ -1,3 +1,8 @@
+# version 1.1.12-preview.1
+## Keyvault exportKey did not add the kid in the jwk.
+**Type of change:** engineering    
+**Customer impact:** low
+
 # version 1.1.12-preview.0
 ## Remove all console.log calls from the SDK
 **Type of change:** engineering